implement results len annotations
