+++ lua/esearch/neovim.lua
@@ -1,6 +1,6 @@
 return {
-  highlight = require'esearch/neovim/highlight',
-  render    = require'esearch/neovim/render'.render,
-  parse     = require'esearch/neovim/parse',
-  util      = require'esearch/util',
+  appearance = require'esearch/neovim/appearance',
+  render     = require'esearch/neovim/render'.render,
+  parse      = require'esearch/neovim/parse',
+  util       = require'esearch/util',
 }
