+++ lua/esearch/neovim/render.lua
@@ -13,6 +13,8 @@ local function render(data, path, last_context, files_count, highlights_enabled)
     vim.api.nvim_get_var('unload_context_syntax_on_line_length')
   local unload_global_syntax_on_line_length =
     vim.api.nvim_get_var('unload_global_syntax_on_line_length')
+  local esearch_win_results_len_annotations =
+    vim.api.nvim_get_var('esearch_win_results_len_annotations')
 
   local start = vim.api.nvim_buf_line_count(0)
   local line = start
@@ -59,7 +61,7 @@ local function render(data, path, last_context, files_count, highlights_enabled)
     end
 
     if text:len() > unload_context_syntax_on_line_length then
-      if text:len() > unload_global_syntax_on_line_length then
+      if text:len() > unload_global_syntax_on_line_length and highlights_enabled == 1 then
         vim.api.nvim_eval('esearch#out#win#stop_highlights()')
       else
         contexts[#contexts]['syntax_loaded'] = -1
@@ -78,8 +80,11 @@ local function render(data, path, last_context, files_count, highlights_enabled)
 
   vim.api.nvim_buf_set_lines(0, -1, -1, 0, lines)
   if vim.api.nvim_eval('g:esearch_out_win_nvim_lua_syntax') == 1 then
-    esearch.highlight.header()
-    esearch.highlight.linenrs_range(0, start, -1)
+    esearch.appearance.header()
+    esearch.appearance.linenrs_range(0, start, -1)
+  end
+  if esearch_win_results_len_annotations == 1 then
+    esearch.appearance.annotate(contexts)
   end
 
   return {files_count, contexts, ctx_ids_map, line_numbers_map, context_by_name, separators_count, highlights_enabled}
