+++ lua/esearch/vim/render.lua
@@ -59,7 +59,7 @@ function render(data, path, esearch)
     end
 
     if text:len() > unload_context_syntax_on_line_length then
-      if text:len() > unload_global_syntax_on_line_length then
+      if text:len() > unload_global_syntax_on_line_length and esearch['highlights_enabled'] then
         vim.eval('esearch#out#win#stop_highlights()')
       else
         contexts[#contexts - 1]['syntax_loaded'] = true
